eliminate format warnings by explicit casting

+++ src/cc65/codegen.c
@@ -4146,15 +4146,15 @@ void g_defdata (unsigned flags, uintptr_t val, long offs)
         switch (flags & CF_TYPEMASK) {
 
             case CF_CHAR:
-                AddDataLine ("\t.byte\t$%02lX", val & 0xFF);
+                AddDataLine ("\t.byte\t$%02lX", (unsigned long)(val & 0xFF));
                 break;
 
             case CF_INT:
-                AddDataLine ("\t.word\t$%04lX", val & 0xFFFF);
+                AddDataLine ("\t.word\t$%04lX", (unsigned long)(val & 0xFFFF));
                 break;
 
             case CF_LONG:
-                AddDataLine ("\t.dword\t$%08lX", val & 0xFFFFFFFF);
+                AddDataLine ("\t.dword\t$%08lX", (unsigned long)(val & 0xFFFFFFFF));
                 break;
 
             default:
